+++ internal/core/usercommands/runtime/runners/host/host.go
@@ -124,6 +124,7 @@ func (r *Runner) Run(ctx context.Context, rc spec.RunContext) error {
 	if err != nil {
 		return err
 	}
+	runio.TraceCommand(ctx, c)
 	defer runio.WireChildIO(rc, c)()
 	return c.Run()
 }