+++ internal/core/usercommands/runtime/runners/service/exec.go
@@ -95,6 +95,7 @@ func (e *ExecRunner) Run(ctx context.Context, rc spec.RunContext) error {
 	if err != nil {
 		return err
 	}
+	runio.TraceCommand(ctx, c)
 	defer runio.WireChildIO(rc, c)()
 	return c.Run()
 }